Remove casts from SyscallReturnCode
The updated version of vmm-sys-util allows us to use different types
with SyscallReturnCode, so many places don't need type conversions for
the return codes anymore.

+++ src/jailer/src/chroot.rs
@@ -83,7 +83,7 @@ pub fn chroot(path: &Path) -> Result<(), JailerError> {
     // SAFETY: Safe because we provide valid parameters.
     SyscallReturnCode(unsafe {
         libc::syscall(libc::SYS_pivot_root, cwd.as_ptr(), old_root_dir.as_ptr())
-    } as libc::c_int)
+    })
     .into_empty_result()
     .map_err(JailerError::PivotRoot)?;
